+++ Know Maps/Know Maps Prod/Model/Protocols/CloudCache.swift
@@ -15,8 +15,8 @@ public protocol CloudCache: AnyObject {
     // Fetching and Caching Methods
     func fetch(url: URL, from cloudService: CloudCacheServiceKey) async throws -> Any
     func fetchGroupedUserCachedRecords(for group: String) async throws -> [UserCachedRecord]
-    func storeUserCachedRecord(for group: String, identity: String, title: String, icons: String, list: String, section: String, rating:Int) async throws -> String
-    func updateUserCachedRecordRating(recordId: String, newRating: Int) async throws
+    func storeUserCachedRecord(for group: String, identity: String, title: String, icons: String, list: String, section: String, rating:Double) async throws -> String
+    func updateUserCachedRecordRating(recordId: String, newRating: Double) async throws
     func deleteUserCachedRecord(for cachedRecord: UserCachedRecord) async throws
     func deleteAllUserCachedRecords(for group: String) async throws -> (saveResults: [CKRecord.ID: Result<CKRecord, Error>], deleteResults: [CKRecord.ID: Result<Void, Error>])
     func deleteAllUserCachedGroups() async throws
